refactor: Move investments page.tsx content to app directory
