+++ app/(dashboard)/dashboard/investments/page.tsx
@@ -1,8 +1,12 @@
+import { cookies } from "next/headers";
+import Image from "next/image";
 import { redirect } from "next/navigation";
 
 import { authOptions } from "@/lib/auth";
 import { getCurrentUser } from "@/lib/session";
-import InvestmentsPage from "@/components/investments/page";
+import { isValidJSONString } from "@/lib/utils";
+import { InvestmentsDashboard } from "@/components/investments/components/investment-dashboard";
+import { accounts, mails } from "@/components/investments/data";
 
 export const metadata = {
   title: "Transactions",
@@ -16,5 +20,47 @@ export default async function DashboardPage() {
     redirect(authOptions?.pages?.signIn || "/login");
   }
 
-  return <InvestmentsPage />;
+  const layout = cookies().get("react-resizable-panels:layout");
+  const collapsed = cookies().get("react-resizable-panels:collapsed");
+
+  //TODO Fix this - Draw bar right is not working. Look at documentation
+
+  const defaultLayout =
+    layout && isValidJSONString(layout.value)
+      ? JSON.parse(layout.value)
+      : undefined;
+  const defaultCollapsed =
+    collapsed && isValidJSONString(collapsed.value)
+      ? JSON.parse(collapsed.value)
+      : undefined;
+
+  return (
+    <>
+      <div className="md:hidden">
+        <Image
+          src="/examples/mail-dark.png"
+          width={1280}
+          height={727}
+          alt="Mail"
+          className="hidden dark:block"
+        />
+        <Image
+          src="/examples/mail-light.png"
+          width={1280}
+          height={727}
+          alt="Mail"
+          className="block dark:hidden"
+        />
+      </div>
+      <div className="hidden flex-col md:flex">
+        <InvestmentsDashboard
+          accounts={accounts}
+          mails={mails}
+          defaultLayout={defaultLayout}
+          defaultCollapsed={defaultCollapsed}
+          navCollapsedSize={4}
+        />
+      </div>
+    </>
+  );
 }
